Keep OCR validation move separate from PaddleOCR
