+++ scinoephile/image/ocr/char_cursor.py
