+++ scinoephile/image/ocr/char_dims.py
