+++ scinoephile/image/ocr/char_grp_dims.py
