+++ scinoephile/image/ocr/char_pair_gaps.py
