+++ scinoephile/image/ocr/gap_cursor.py
