--- scinoephile/image/ocr/validation/__init__.py
@@ -1,9 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""OCR validation for image subtitles."""
-
-from __future__ import annotations
-
-from .validation_manager import ValidationManager
-
-__all__ = ["ValidationManager"]
