+++ scinoephile/image/ocr/validation_manager.py
